Extract inner classes
